make gcc and sparse happy (mass commit)
This one deals with undefined functions coming from neighbor modules
which are only recognized during linking, non-ansi function definitions,
const chars vs chars and all sorts of things which are easily spotted
by a C compiler front-end when it's asked to do so.

Note however, this is only testsuite-verified and not absolutely
guaranteed to work on real-world stuff (but is pretty much likely
to).

Signed-off-by: Alexander Shishkin <virtuoso@slind.org>