add more generic parser for control files
This commit adds parse_buf() function (along with certain supporting
calls) which, given a buffer containing a control file (either from
a binary, or a source package) breaks it into key/value pairs which
in turn are fed to a callback.

This also eliminates two dpkg-deb calls for each binary package in
a repository.

Signed-off-by: Alexander Shishkin <virtuoso@slind.org>